Finished Cryptocurrency class. Working on Wallet and Trade. Added default returns for all nonimplemented classes. Fixed a bunch of errors and updated a few types.

+++ CryptoProject/src/Client/Currency.java
@@ -5,11 +5,9 @@ This class holds the basic object which is later derived from to create Cryptocu
 of how derivation works in OOP.
 */
 
-import java.math.BigDecimal;
-
 abstract class Currency {
 
     String name;
-    BigDecimal amountHeld;
+    Double amountHeld;
 
 }
